[OpenCL] Initial attempt at splitting Conv2D
By splitting the convolution kernel into two, first expanding the input
tensor and then performing the contraction, we can improve the memory
accesses in the contraction kernel and improve performance. However the
introduction of a large intermediate tensor can cause problems if there
is a limit to how much memory can be allocated.