[SNN] Splits kernel execution for large tensors
For large tensors, the intermediate tensors required for im2col and
winograd convolutions can be too large to allocate on the device. We try
to split these convolutions up, essentially decreasing the size of the
batches to allow the tensors to be allocated. If doing this is not
enough to give sufficiently small buffers we fall back on using the
direct convolutions which do not require any allocations.